// src/app/core/models/user/owner-profile.model.ts
export interface OwnerProfile {
  cover: string;
  following: string[];
  _id: string;
  name: string;
  email: string;
  dateOfBirth: string;
  gender: string;
  photo: string;
  createdAt: string;
  followers: string[];
  bookmarks: string[];
  followersCount: number;
  followingCount: number;
  bookmarksCount: number;
  id: string;
}